+++ parsley-core/src/ghc/Parsley/Internal.hs
@@ -16,11 +16,26 @@ module Parsley.Internal (
     module Primitives,
     module THUtils,
     module Trace,
+#if MIN_VERSION_parsley_core(2,0,0)
+#else
     module Frontend,
-    module Backend
+#endif
+    module Backend,
+    parse
   ) where
 
-import Parsley.Internal.Backend         as Backend    (codeGen, Input, eval)
+#if MIN_VERSION_parsley_core(2,0,0)
+import Parsley.Internal.Backend  (codeGen, eval)
+import Parsley.Internal.Frontend (compile)
+#endif
+
+import Parsley.Internal.Backend         as Backend    (
+    Input,
+#if MIN_VERSION_parsley_core(2,0,0)
+#else
+    codeGen, eval
+#endif
+  )
 import Parsley.Internal.Core            as Core
 import Parsley.Internal.Core.Primitives as Primitives (
     pure, (<*>), (*>), (<*),
@@ -36,5 +51,11 @@ import Parsley.Internal.Core.Primitives as Primitives (
     debug
   )
 import Parsley.Internal.Common.Utils    as THUtils    (Quapplicative(..), WQ, Code, makeQ)
+#if MIN_VERSION_parsley_core(2,0,0)
+#else
 import Parsley.Internal.Frontend        as Frontend   (compile)
+#endif
 import Parsley.Internal.Trace           as Trace      (Trace(trace))
+
+parse :: (Trace, Input input) => Parser a -> Code (input -> Maybe a)
+parse p = [||\input -> $$(eval [||input||] (compile (try p) codeGen))||]