+++ parsley/src/ghc/Parsley.hs
@@ -1,5 +1,5 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE MultiParamTypeClasses, CPP #-}
 {-|
 Module      : Parsley
 Description : Main functionality exports
@@ -29,7 +29,7 @@ module Parsley (
 import Prelude hiding            (readFile)
 import Data.Text.IO              (readFile)
 import Parsley.InputExtras       (Text16(..))
-import Parsley.Internal          (codeGen, Input, eval, compile, Trace(trace))
+import Parsley.Internal          (Input, Trace(trace))
 
 import Parsley.Alternative              as Alternative
 import Parsley.Applicative              as Applicative
@@ -40,6 +40,12 @@ import Parsley.ParserOps                as Core        (ParserOps)
 import Parsley.Internal                 as THUtils     (Quapplicative(..), WQ, Code)
 import Parsley.Debug                    as Primitives  (debug)
 import Parsley.Selective                as Selective
+
+#if MIN_VERSION_parsley_core(1,7,1)
+import qualified Parsley.Internal as Internal (parse)
+#else
+import qualified Parsley.Internal as Internal (eval, compile, codeGen)
+#endif
 
 {-|
 The standard way to compile a parser, it returns `Code`, which means
@@ -69,7 +75,11 @@ the compilation pipeline when "Parsley.Internal.Verbose" is imported.
 runParser :: (Trace, Input input)
           => Parser a                -- ^ The parser to be compiled
           -> Code (input -> Maybe a) -- ^ The generated parsing function
-runParser p = [||\input -> $$(eval [||input||] (compile (try p) codeGen))||]
+#if MIN_VERSION_parsley_core(1,7,1)
+runParser = Internal.parse
+#else
+runParser p = [||\input -> $$(Internal.eval [||input||] (Internal.compile (try p) Internal.codeGen))||]
+#endif
 
 {-|
 This function generates a function that reads input from a file